Fix: directory-cleanup
